+++ src/game/services/interfaces/time_service.hpp
@@ -10,8 +10,8 @@ namespace Game {
             public:
                 virtual ~TimeService() {}
                 virtual void BeginFrame() = 0;
-                virtual double GetDeltaTime() const = 0;
-                virtual double GetElapsedTime() const = 0;
+                virtual float GetDeltaTime() const = 0;
+                virtual float GetElapsedTime() const = 0;
                 virtual void IncrementFrameCount() = 0;
                 virtual void StartClock(const unsigned int offset) = 0;
                 virtual unsigned int GetFrameCount() = 0;
