+++ src/game/states/battle.cpp
@@ -43,7 +43,7 @@ namespace Game {
             }
         }
 
-        Transition Battle::Update(const double deltaTime) {
+        Transition Battle::Update(const float deltaTime) {
             std::string nextState = (*this->luaState.get())["update"](deltaTime);
 
             if (this->pop) {
