+++ src/game/states/intro.hpp
@@ -13,7 +13,7 @@ namespace Game {
             Intro();
             ~Intro();
             void HandleEvent(const SDL_Event& event);
-            Transition Update(const double deltaTime);
+            Transition Update(const float deltaTime);
             void Render();
 
         private:
