+++ src/game/states/main_menu.hpp
@@ -13,7 +13,7 @@ namespace Game {
             MainMenu();
             ~MainMenu();
             void HandleEvent(const SDL_Event& event);
-            Transition Update(const double deltaTime);
+            Transition Update(const float deltaTime);
             void Render();
 
         private:
