+++ src/game/states/pause_menu.hpp
@@ -16,7 +16,7 @@ namespace Game {
             PauseMenu();
             ~PauseMenu();
             void HandleEvent(const SDL_Event& event);
-            Transition Update(const double deltaTime);
+            Transition Update(const float deltaTime);
             void Render();
             Objects::Characters::Party GetParty();
             std::vector<std::shared_ptr<Objects::Characters::PlayerCharacter>> GetPartyCharacters();
